Add consent integration test utilities and coverage
