+++ MonoKnightAppTests/AdsConsentCoordinatorTests.swift
@@ -1,87 +1,6 @@
-import SwiftUI
-import UIKit
 import Testing
 import UserMessagingPlatform
 @testable import MonoKnightApp
-
-// MARK: - スタブ定義
-@MainActor
-private final class StubAdsConsentEnvironment: AdsConsentEnvironment {
-    /// UMP 側の consentStatus をテストから書き換えやすく保持
-    var consentStatus: ConsentStatus = .unknown
-    /// フォームの利用可否をテストシナリオ毎に指定
-    var formStatus: FormStatus = .unknown
-    /// canRequestAds を外部から操作し、ロード条件を切り替える
-    var canRequestAds: Bool = false
-    /// requestConsentInfoUpdate が呼ばれた回数
-    private(set) var requestUpdateCallCount: Int = 0
-    /// loadConsentFormPresenter が呼ばれた回数
-    private(set) var loadFormCallCount: Int = 0
-    /// makePrivacyOptionsPresenter が呼ばれた回数
-    private(set) var makePrivacyOptionsCallCount: Int = 0
-
-    /// 更新時に追加で実行したい処理（例: consentStatus の差し替え）
-    var requestUpdateHandler: (() -> Void)?
-    /// フォーム表示用クロージャを差し替えるためのプロパティ
-    var presenterFactory: (() -> ConsentFormPresenter)?
-    /// プライバシーオプション表示用クロージャを差し替えるためのプロパティ
-    var privacyPresenterFactory: (() -> PrivacyOptionsPresenter)?
-
-    func requestConsentInfoUpdate(with parameters: RequestParameters) async throws {
-        requestUpdateCallCount += 1
-        requestUpdateHandler?()
-    }
-
-    func loadConsentFormPresenter() async throws -> ConsentFormPresenter {
-        loadFormCallCount += 1
-        if let presenterFactory {
-            return presenterFactory()
-        }
-        return { _, completion in completion(nil) }
-    }
-
-    func makePrivacyOptionsPresenter() -> PrivacyOptionsPresenter {
-        makePrivacyOptionsCallCount += 1
-        if let privacyPresenterFactory {
-            return privacyPresenterFactory()
-        }
-        return { _, completion in completion(nil) }
-    }
-}
-
-@MainActor
-private final class StubConsentPresentationDelegate: AdsConsentCoordinatorPresenting {
-    /// 同意フォーム表示要求の回数
-    private(set) var presentConsentFormCallCount: Int = 0
-    /// プライバシーオプション表示要求の回数
-    private(set) var presentPrivacyOptionsCallCount: Int = 0
-    /// 直近で受け取った presenter（検証用）
-    private(set) var lastConsentPresenter: ConsentFormPresenter?
-    /// 直近で受け取ったプライバシー presenter
-    private(set) var lastPrivacyPresenter: PrivacyOptionsPresenter?
-
-    func presentConsentForm(using presenter: @escaping ConsentFormPresenter) async throws {
-        presentConsentFormCallCount += 1
-        lastConsentPresenter = presenter
-        presenter(UIViewController()) { _ in }
-    }
-
-    func presentPrivacyOptions(using presenter: @escaping PrivacyOptionsPresenter) async throws {
-        presentPrivacyOptionsCallCount += 1
-        lastPrivacyPresenter = presenter
-        presenter(UIViewController()) { _ in }
-    }
-}
-
-@MainActor
-private final class StubConsentStateDelegate: AdsConsentCoordinatorStateDelegate {
-    /// 状態更新の履歴（shouldReload フラグも保持）
-    private(set) var recordedStates: [(state: AdsConsentState, shouldReload: Bool)] = []
-
-    func adsConsentCoordinator(_ coordinator: AdsConsentCoordinating, didUpdate state: AdsConsentState, shouldReloadAds: Bool) {
-        recordedStates.append((state, shouldReloadAds))
-    }
-}
 
 // MARK: - テスト本体
 struct AdsConsentCoordinatorTests {
@@ -90,7 +9,7 @@ struct AdsConsentCoordinatorTests {
     @Test func requestConsentIfNeeded_showsFormWhenRequired() async throws {
         UserDefaults.standard.removeObject(forKey: "ads_should_use_npa")
 
-        let environment = StubAdsConsentEnvironment()
+        let environment = TestAdsConsentEnvironment()
         environment.consentStatus = .required
         environment.formStatus = .available
         environment.canRequestAds = false
@@ -105,8 +24,8 @@ struct AdsConsentCoordinatorTests {
             }
         }
 
-        let presenter = StubConsentPresentationDelegate()
-        let stateDelegate = StubConsentStateDelegate()
+        let presenter = TestConsentPresentationDelegate()
+        let stateDelegate = TestConsentStateRecorder()
         let coordinator = AdsConsentCoordinator(hasValidAdConfiguration: true, environment: environment)
         coordinator.presentationDelegate = presenter
         coordinator.stateDelegate = stateDelegate
@@ -130,7 +49,7 @@ struct AdsConsentCoordinatorTests {
     @Test func refreshConsentStatus_showsPrivacyOptions() async throws {
         UserDefaults.standard.removeObject(forKey: "ads_should_use_npa")
 
-        let environment = StubAdsConsentEnvironment()
+        let environment = TestAdsConsentEnvironment()
         environment.consentStatus = .obtained
         environment.formStatus = .available
         environment.canRequestAds = true
@@ -142,8 +61,8 @@ struct AdsConsentCoordinatorTests {
             }
         }
 
-        let presenter = StubConsentPresentationDelegate()
-        let stateDelegate = StubConsentStateDelegate()
+        let presenter = TestConsentPresentationDelegate()
+        let stateDelegate = TestConsentStateRecorder()
         let coordinator = AdsConsentCoordinator(hasValidAdConfiguration: true, environment: environment)
         coordinator.presentationDelegate = presenter
         coordinator.stateDelegate = stateDelegate
@@ -159,9 +78,9 @@ struct AdsConsentCoordinatorTests {
     /// Info.plist の設定が不完全な場合、環境側の処理が呼ばれないことを保証
     @MainActor
     @Test func coordinator_skipsOperations_whenConfigurationIsInvalid() async throws {
-        let environment = StubAdsConsentEnvironment()
-        let presenter = StubConsentPresentationDelegate()
-        let stateDelegate = StubConsentStateDelegate()
+        let environment = TestAdsConsentEnvironment()
+        let presenter = TestConsentPresentationDelegate()
+        let stateDelegate = TestConsentStateRecorder()
         let coordinator = AdsConsentCoordinator(hasValidAdConfiguration: false, environment: environment)
         coordinator.presentationDelegate = presenter
         coordinator.stateDelegate = stateDelegate
